add mock_test.go
